client: remove, temporarily, all activation functionality
It turns out it isn't necessary to configure and start a `LanguageClient` object
for (client-side) syntax highlighting to take effect. Because we did so by
referring to a nonexistent server executable, the client would always display a
needless error message on startup. Getting rid of this code is the cleanest way
to get rid of the error.